Update dependencies on 'dev'
Also, remove optional dependency to websockets, which is not
part of Qt 6.0 and fix compile error against unsigned comparison.

Switch default test configuration to not launch TLS brokers.
Latest COIN updates causes paho to fail in the SSL setup.

Task-number: QTBUG-88189

+++ src/mqtt/qmqttpublishproperties.cpp
@@ -244,7 +244,7 @@ QList<quint32> QMqttPublishProperties::subscriptionIdentifiers() const
 */
 void QMqttPublishProperties::setSubscriptionIdentifiers(const QList<quint32> &ids)
 {
-    if (ids.contains(0)) {
+    if (ids.contains(quint32(0))) {
         qCDebug(lcMqttClient) << "A subscription identifier with value 0 is not allowed.";
         return;
     }
